fix: error handlign for setting client
